Refs #16. Reorganize codes.
