+++ src/ui/AppLogManager.py
