+++ src/ui/LogSnapView.py
@@ -14,7 +14,7 @@ except AttributeError:
         return s
         
 import GuiUtility as gutil
-import ui_LogSnapView
+import gui.ui_LogSnapView as ui_LogSnapView
 
 
 class DialogLogSnapView(QtGui.QDialog):
