+++ src/ui/gui/Makefile
@@ -1,0 +1,15 @@
+#
+# Make file to
+#
+#
+UIFILES := $(wildcard *.ui)
+PYFILES := $(UIFILES:.ui=.py)
+
+.PHONY: all
+all: $(PYFILES)
+
+%.py: %.ui
+	pyuic4 $< --output $@
+
+clean: 
+	rm *pyc */*.pyc
