+++ src/ui/gui/dataslicersegmenttable.py
