+++ src/ui/gui/ui_AppLog.ui
