+++ src/ui/gui/ui_LogSnapView.ui
